requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
starlette==0.46.1
SQLAlchemy==2.0.39
alembic==1.15.1
asyncpg==0.30.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.4.0
pydantic==2.10.6
pydantic-settings==2.8.1
fastapi-mail==1.4.2
slowapi==0.1.9
cloudinary==1.43.0
libgravatar==1.0.4
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
aioredis==2.0.1
aiosqlite==0.21.0
alabaster==1.0.0
async-timeout==5.0.1
babel==2.17.0
charset-normalizer==3.4.1
coverage==7.7.1
docutils==0.21.2
httpcore==1.0.7
httpx==0.28.1
imagesize==1.4.1
iniconfig==2.1.0
pluggy==1.5.0
Pygments==2.19.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
redis==5.2.1
redis-lru==0.1.2
requests==2.32.3
roman-numerals-py==3.1.0
snowballstemmer==2.2.0
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib.httpdomain==1.8.0